Covert clair support to a plugin
