+++ .env.example
@@ -113,3 +113,6 @@ GITHUB_TOKEN=
 # JENKINS_URL= # server_url of jenkins
 # JENKINS_USERNAME= # user id
 # JENKINS_API_KEY= # API Token from user / Configure page
+
+## Hyperclair, optional to security scan built docker images
+# HYPERCLAIR_PATH=/usr/local/bin/hyperclair
