+++ plugins/hyperclair/samson_hyperclair.gemspec
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+Gem::Specification.new "samson_hyperclair", "0.0.0" do |s|
+  s.summary = "Samson Hyperclair integration"
+  s.authors = ["Michael Grosser"]
+  s.email = "mgrosser@zendesk.com"
+end
