+++ plugins/hyperclair/test/test_helper.rb
@@ -1,0 +1,2 @@
+# frozen_string_literal: true
+require_relative '../../../test/test_helper'
